requirements.txt
python-dotenv
google-adk
google-genai
uv
yfinance